Modifying get call to native LLVM IR
Generics have to be implemented in the LLVM IR itself, as ABI functions
are designed to only accept specific types and it's not feasible
to specialize for all known (and declared) types.

There is also a performance benefit. With 500k iterations:

* using function call: 0.0068
* llvm IR: 0.000814277

Roughly a 8-9x speedup.

One gotcha was in the construction fo the element pointer: GEP
 instructions do not dereference pointers and instead only
do pointer arithmetic, so the Load instruction must be used
to deference pointers before subsequent GEP calls.

There is also a change to move all integers to i32. This is temporary but helps
avoid an issue where GEP instructions for structs can only consume i32 values.

+++ src/expressions/bytes_builtins.rs
@@ -4,7 +4,7 @@ use std::{slice, str};
 type Bytes = Array<u8>;
 
 #[no_mangle]
-pub extern "C" fn get_bytes(bytes: *mut Bytes, index: i64) -> u8 {
+pub extern "C" fn get_bytes(bytes: *mut Bytes, index: i32) -> u8 {
     unsafe { slice::from_raw_parts((*bytes).values, (*bytes).size as usize)[index as usize] }
 }
 
@@ -22,6 +22,6 @@ pub extern "C" fn print_byte(value: u8) {
 }
 
 #[no_mangle]
-pub extern "C" fn len_bytes(bytes: *mut Bytes) -> i64 {
+pub extern "C" fn len_bytes(bytes: *mut Bytes) -> i32 {
     unsafe { (*bytes).size }
 }
